Date is set on creation
Date gets set on ticket creation, but it is not seen when viewing the ticket as agent+

Date is seen in ticket list but not in the single view. 

Field still expects an input and will save a blank date if the agent+ applies any updates without setting the correct due date.